refactor: 优化Black Jack游戏代码
